Move GpuParquetScan to shim
